backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]==2.10.4
httpx==0.28.1
redis==5.2.1
orjson==3.10.12
python-dateutil==2.9.0.post0
lxml==5.3.0

SQLAlchemy==2.0.36
geoalchemy2==0.18.1
asyncpg==0.31.0
alembic==1.14.0
psycopg2-binary==2.9.11

celery==5.4.0
kombu==5.4.2

tenacity==9.0.0

openmeteo-requests==1.7.4
requests-cache==1.2.1
retry-requests==2.0.0
requests==2.32.5
joblib==1.5.3

pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0
xgboost==3.1.2
lightgbm==4.6.0

python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.21